requirements.txt
Flask==3.0.3
google-generativeai==0.8.1
requests==2.32.3